Améliorations des fonctionnalités
- Message de confirmation avant la suppression de compte
- Possibilité pour les admins de modifier les profils (sans entrer le mot de passe
- Accès aux publications pour un utilisateur spécifique ouvert à tous les utilisateurs
- Amélioration de la présentation des posts et commentaires (possibilité de cliquer sur le pseudo de l'auteur pour accéder à son profil)

+++ frontend/src/components/User/DeleteUser.vue
@@ -1,16 +1,35 @@
 <template>
   <div>
-    <button
-      type="submit"
-      class="btn btn-danger button"
-      @click="deleteAccount()"
+    <b-button v-b-modal.delete-account-modal class="btn btn-danger button"
+      >Supprimer le compte</b-button
     >
-      <h2 class="">
-        <b-icon-exclamation-triangle-fill></b-icon-exclamation-triangle-fill>
-        Supprimer le compte {{ user.username }}
-        <b-icon-exclamation-triangle-fill></b-icon-exclamation-triangle-fill>
-      </h2>
-    </button>
+
+    <b-modal hide-footer id="delete-account-modal" title="Supression du compte">
+      <p id="confirmation-text">
+        En cliquant sur "supprimer", le compte et toutes les informations
+        relatives au compte seront supprimées. Souhaitez-vous vraiment
+        continuer?
+      </p>
+
+      <button
+        type="submit"
+        class="btn btn-danger button"
+        @click="checkIdentity()"
+      >
+        test
+      </button>
+      <button
+        type="submit"
+        class="btn btn-danger button"
+        @click="deleteAccount()"
+      >
+        <h2 class="">
+          <b-icon-exclamation-triangle-fill></b-icon-exclamation-triangle-fill>
+          Supprimer le compte {{ user.username }}
+          <b-icon-exclamation-triangle-fill></b-icon-exclamation-triangle-fill>
+        </h2>
+      </button>
+    </b-modal>
   </div>
 </template>
 
@@ -23,23 +42,48 @@ let apiUrl = "http://localhost:" + apiPort + "/api/";
 export default {
   data() {
     return {
-      user: "",
       routeUserId: this.$route.params.userId,
+      userIsAdmin: false,
+      userIsOwner: false,
     };
   },
+  props: ["user", "routeUser"],
   methods: {
+    checkIdentity() {
+      if (this.routeUserId == this.user.userId) {
+        this.userIsOwner = true;
+        console.log("user is owner");
+      } else {
+        this.userIsOwner = false;
+      }
+      if (this.user.roleId == 1) {
+        this.userIsAdmin = true;
+        console.log("user is admin");
+      } else {
+        this.userIsAdmin = false;
+      }
+    },
     deleteAccount() {
-      console.log("User : " + this.user);
+      this.checkIdentity(); // On vérifie que l'utilisateur soit ou l'admin ou le propriétaire de la page, ou les deux
+      console.log("User : " + this.routeUserId);
       axios
         .delete(apiUrl + "auth/users/" + this.routeUserId, {
           headers: { Authorization: "Bearer " + localStorage.token },
         })
         .then((response) => {
           console.log(response);
-          localStorage.clear();
-          this.$store.dispatch("user/logout").then(() => {
-            this.$router.push("/Auth");
-          });
+          if (this.userIsAdmin && !this.userIsOwner) {
+            // Si l'utilisateur est admin mais pas propriétaire de la page, alors on le redirige vers la homepage du site après la suppression
+            this.$router.push("/");
+          } else if (
+            (this.userIsOwner && this.userIsAdmin) ||
+            (this.userIsOwner && !this.userIsAdmin)
+          ) {
+            // Si l'utilisateur est admin et propriétaire de la page ou pas admin mais propriétaire de la page, alors on le redirige vers la page de connexion après avoir supprimé le localStorage et les informations de connexion
+            localStorage.clear();
+            this.$store.dispatch("user/logout");
+            this.$router.push("/");
+          }
         })
         .catch((err) => console.log(err));
     },
